refactor: move app and config to the core directory
